+++ Sources/SwifterSwiftUI/Foundation/Optional+Extensions.swift
@@ -1,5 +1,5 @@
 //
-//  Optional.swift
+//  Optional+Extensions.swift
 //  SwifterSwiftUI
 //
 //  Created by Saul Moreno Abril on 08/06/2020.